# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "ApiResponse_null_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "anyOf": [],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "CastData": {
        "properties": {
          "name": {
            "type": "string"
          },
          "character": {
            "type": "string"
          },
          "profileImage": {
            "type": "string"
          }
        },
        "required": ["name", "character", "profileImage"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Series": {
        "properties": {
          "id": {
            "type": "string"
          },
          "libraryId": {
            "type": "string"
          },
          "themdbId": {
            "type": "number",
            "format": "double"
          },
          "order": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "nameLock": {
            "type": "boolean"
          },
          "overview": {
            "type": "string"
          },
          "overviewLock": {
            "type": "boolean"
          },
          "year": {
            "type": "string"
          },
          "yearLock": {
            "type": "boolean"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "tagline": {
            "type": "string"
          },
          "taglineLock": {
            "type": "boolean"
          },
          "logoSrc": {
            "type": "string"
          },
          "logosUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "coverSrc": {
            "type": "string"
          },
          "coversUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "productionStudios": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "productionStudiosLock": {
            "type": "boolean"
          },
          "creator": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "creatorLock": {
            "type": "boolean"
          },
          "musicComposer": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "musicComposerLock": {
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "genresLock": {
            "type": "boolean"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastData"
            },
            "type": "array"
          },
          "preferAudioLan": {
            "type": "string"
          },
          "preferSubLan": {
            "type": "string"
          },
          "subsMode": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "episodeGroupId": {
            "type": "string",
            "nullable": true
          },
          "analyzingFiles": {
            "type": "boolean"
          },
          "watchLists": {
            "items": {
              "$ref": "#/components/schemas/WatchList"
            },
            "type": "array"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/Season"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "libraryId",
          "themdbId",
          "order",
          "name",
          "nameLock",
          "overview",
          "overviewLock",
          "year",
          "yearLock",
          "score",
          "tagline",
          "taglineLock",
          "logoSrc",
          "logosUrls",
          "coverSrc",
          "coversUrls",
          "productionStudios",
          "productionStudiosLock",
          "creator",
          "creatorLock",
          "musicComposer",
          "musicComposerLock",
          "genres",
          "genresLock",
          "cast",
          "folder",
          "episodeGroupId",
          "analyzingFiles",
          "watchLists",
          "seasons"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetVideoWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetMovieWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          }
        },
        "required": ["watched"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ApiResponse_number_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "RefreshMetadataDTO": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateShowIdDTO": {
        "properties": {
          "id": {
            "type": "string"
          },
          "themdbId": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["id", "themdbId"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateEpisodeGroupDTO": {
        "properties": {
          "id": {
            "type": "string"
          },
          "themdbId": {
            "type": "number",
            "format": "double"
          },
          "episodeGroupId": {
            "type": "string"
          }
        },
        "required": ["id", "themdbId", "episodeGroupId"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_Series_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Series"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateSeriesDTO": {
        "properties": {
          "name": {
            "type": "string"
          },
          "nameLock": {
            "type": "boolean"
          },
          "overview": {
            "type": "string"
          },
          "overviewLock": {
            "type": "boolean"
          },
          "year": {
            "type": "string"
          },
          "yearLock": {
            "type": "boolean"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "tagline": {
            "type": "string"
          },
          "taglineLock": {
            "type": "boolean"
          },
          "logoSrc": {
            "type": "string"
          },
          "logosUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "coverSrc": {
            "type": "string"
          },
          "coversUrls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "productionStudios": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "productionStudiosLock": {
            "type": "boolean"
          },
          "creator": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "creatorLock": {
            "type": "boolean"
          },
          "musicComposer": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "musicComposerLock": {
            "type": "boolean"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "genresLock": {
            "type": "boolean"
          },
          "preferAudioLan": {
            "type": "string"
          },
          "preferSubLan": {
            "type": "string"
          },
          "subsMode": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "episodeGroupId": {
            "type": "string",
            "nullable": true
          },
          "analyzingFiles": {
            "type": "boolean"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "SetSeriesWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
      "IncludeType": {
        "type": "string",
        "enum": ["none", "few", "all"]
      },
      "TvResult": {
        "properties": {
          "poster_path": {
            "type": "string"
          },
          "popularity": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "number",
            "format": "double"
          },
          "overview": {
            "type": "string"
          },
          "backdrop_path": {
            "type": "string"
          },
          "vote_average": {
            "type": "number",
            "format": "double"
          },
          "media_type": {
            "type": "string",
            "enum": ["tv"],
            "nullable": false
          },
          "first_air_date": {
            "type": "string"
          },
          "origin_country": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "genre_ids": {
            "items": {
              "type": "number",
              "format": "double"
            },
            "type": "array"
          },
          "original_language": {
            "type": "string"
          },
          "vote_count": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "original_name": {
            "type": "string"
          }
        },
        "required": ["media_type"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_TvResult-Array_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TvResult"
            },
            "type": "array",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ApiResponse_TvEpisodeGroupsResponse-or-null_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TvEpisodeGroupsResponse"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetSeasonWatchStateDTO": {
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": ["watched", "userId"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetEpisodeWatchStateDTO": {
        "properties": {
          "state": {
            "type": "boolean"
          }
        },
        "required": ["state"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Full user authentication with library and remote access checks"
      },
      "cookieAuthFast": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Fast user authentication without additional checks"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Admin user authentication"
      },
      "managementAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Management access for local or admin users"
      },
      "public": {
        "type": "apiKey",
        "in": "header",
        "name": "x-public-access",
        "description": "Public access without authentication"
      }
    }
  },
  "info": {
    "title": "Seerial Media Server API",
    "version": "0.2.0",
    "description": "Media management server API for the Seerial suite",
    "license": {
      "name": "GPL-3.0-only"
    },
    "contact": {
      "name": "Juan Llavero"
    }
  },
  "paths": {
...
    "/videos/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set video watch state for a user",
        "tags": ["Videos"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVideoWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/movies/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set movie watch state for a user",
        "tags": ["Movies"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMovieWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/series/metadata": {
      "post": {
        "operationId": "RefreshMetadata",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Refresh series metadata from external sources",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshMetadataDTO"
              }
            }
          }
        }
      }
    },
    "/series/tmdb-id": {
      "post": {
        "operationId": "UpdateShowId",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Update series TMDB ID",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShowIdDTO"
              }
            }
          }
        }
      }
    },
    "/series/{id}/episode-group": {
      "post": {
        "operationId": "UpdateEpisodeGroup",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Update series episode group",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEpisodeGroupDTO"
              }
            }
          }
        }
      }
    },
    "/series/show/{id}": {
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Series_"
                }
              }
            }
          }
        },
        "description": "Update series details",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSeriesDTO"
              }
            }
          }
        }
      }
    },
    "/series/{id}": {
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Delete a series",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "get": {
        "operationId": "Get",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Series_"
                }
              }
            }
          }
        },
        "description": "Get series by ID",
        "tags": ["Series"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IncludeType"
            }
          }
        ]
      }
    },
    "/series/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set series watch state for a user",
        "tags": ["Series"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeriesWatchStateDTO"
              }
            }
          }
        }
      }
    },
    "/series/search": {
      "get": {
        "operationId": "SearchSeries",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TvResult-Array_"
                }
              }
            }
          }
        },
        "description": "Search series in TMDB",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/series/episode-groups/search": {
      "get": {
        "operationId": "SearchEpisodeGroups",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TvEpisodeGroupsResponse-or-null_"
                }
              }
            }
          }
        },
        "description": "Search episode groups in TMDB",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/series/{id}/remaining-episodes": {
      "get": {
        "operationId": "GetRemainingEpisodes",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_number_"
                }
              }
            }
          }
        },
        "description": "Get remaining episodes count for a series",
        "tags": ["Series"],
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/seasons/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set season watch state for a user",
        "tags": ["Seasons"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeasonWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
    "/episodes/{id}/watch-state": {
      "post": {
        "operationId": "SetWatchState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Set episode watch state for a user",
        "tags": ["Episodes"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEpisodeWatchStateDTO"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:34200/api"
    }
  ]
}